Final Code, Simplified
This is essentially the final code used for the installation. There was
a video feedback element as the camera was pointed at a projection. The
x value may have been reversed to account for this, I'm not sure. This
project is MIT open source so feel free to use this code!